requirements.txt
polars==1.19.0
pendulum==3.0.0
sqlmodel==0.0.22
numpy==2.2.1
python-dotenv==1.0.1
gspread==6.1.4
oauth2client==4.1.3
google-api-python-client==2.159.0